refactor(core): extract element/bounds selectors into selectors.ts
Pull the next batch of pure derivations out of init() into a new
`createGrabElementSelectors` factory:

- isRendererActive
- targetElement
- effectiveElement
- selectionElement + isSelectionElementVisible (handles touch-mode + drag fallback)
- frozenElementBoundsAccessors
- frozenElementsBounds
- selectionBounds

These selectors depend on the phase selectors plus the store, with no
local-signal coupling, so they're a clean extraction. The previous
local-let-rec patterns (selectionElement reading effectiveElement,
selectionBounds reading frozenElementsBounds, etc.) are preserved
through composition inside the factory.

index.tsx: 3724 -> 3623 lines (-101). No behavior change.
